third_party/WebKit/LayoutTests/inspector/profiler/cpu-profiler-stopped-removed-race-expected.txt

Tests that profile removal right after stop profiling issued works. Bug 476430.

Profiler was enabled.

Running: testProfiling
Stop recording command issued.
resumed
Recording stopped. profile (should be null): null

Profiler was disabled.
